% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{a_j_round}
\alias{a_j_round}
\title{Inverse rounding function: usual rounding + bounds}
\usage{
a_j_round(j, y_min = -Inf, y_max = Inf)
}
\arguments{
\item{j}{the integer-valued input(s)}

\item{y_min}{a fixed and known lower bound for all observations; default is \code{-Inf}}

\item{y_max}{a fixed and known upper bound for all observations; default is \code{Inf}}
}
\value{
The (lower) interval endpoint(s) associated with \code{j}.
}
\description{
Define the intervals associated with \code{y = j} based on the midpoint rounding.
The function returns \code{-Inf} for \code{j < y_min} and \code{Inf} for
 \code{j > y_max}, where \code{y_min} and \code{y_max} are a known bounds on the data \code{y}
(if specified). Negative values are allowed.
}
\examples{
# Standard cases:
a_j_round(0)
a_j_round(20)

# Boundary cases:
a_j_round(0, y_min = 1)
a_j_round(20.5, y_max = 15)

}
