% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{blm_star_exact_bnp}
\alias{blm_star_exact_bnp}
\title{Monte Carlo sampler for STAR linear regression with BNP transformation}
\usage{
blm_star_exact_bnp(
  y,
  X,
  X_test = X,
  y_max = Inf,
  psi = length(y),
  alpha = 1,
  P0 = NULL,
  pilot_run = FALSE,
  nsave = 1000,
  verbose = TRUE
)
}
\arguments{
\item{y}{\code{n x 1} vector of observed counts}

\item{X}{\code{n x p} matrix of predictors (including an intercept)}

\item{X_test}{\code{n_test x p} matrix of predictors for test data
(including an intercept); default is the observed covariates \code{X}}

\item{y_max}{a fixed and known upper bound for all observations; default is \code{Inf}}

\item{psi}{prior variance (g-prior); default is \code{n}}

\item{alpha}{prior precision for the Dirichlet Process prior; default is one}

\item{P0}{function to evaluate the base measure PMF supported on \code{{0,...,y_max}};
see below for default values when unspecified (\code{NULL})}

\item{pilot_run}{logical; if \code{TRUE}, use a short pilot run to approximate
the marginal CDF of the latent \code{z}; otherwise, use a Laplace approximation}

\item{nsave}{number of Monte Carlo iterations to save}
}
\value{
a list with the following elements:
\itemize{
\item \code{coefficients}: the posterior mean of the regression coefficients
\item \code{post.beta}: \code{nsave x p} samples from the posterior distribution
of the regression coefficients
\item \code{post.pred}: \code{nsave x n_test} samples
from the posterior predictive distribution at test points \code{X_test}
\item \code{post.g}: \code{nsave} posterior samples of the transformation
evaluated at \code{1:max(y)}
\item \code{post.log.like.point}: draws of the log-likelihood for each of the \code{n} observations
\item \code{WAIC}: Widely-Applicable/Watanabe-Akaike Information Criterion
\item \code{p_waic}: Effective number of parameters based on WAIC
}
}
\description{
Compute direct Monte Carlo samples from the posterior and predictive
distributions of a STAR linear regression model with a g-prior
and Bayesian nonparametric (BNP) transformation.
}
\details{
This function provides fully Bayesian inference for a
transformed linear model with discrete data. The sampling algorithm draws the
transformation \code{g} together with the regression coefficients \code{beta}
from their *joint* posterior distribution using Monte Carlo (not MCMC) sampling.
When \code{n} is moderate to larger, \link{blm_star_gibbs_bnp} is recommended.

The BNP model for the transformation \code{g} is derived from a Dirichlet Process (DP)
prior on the marginal CDF of \code{y}. The user can specify the prior precision
\code{alpha} and the base measure, e.g., \code{P0 = function(t) dpois(t, lambda = 10)}.
Otherwise, the default approach views the base measure as a means to ensure
correct support and avoid boundary issues, both of which are concerns with the
Bayesian bootstrap (\code{alpha = 0}). Thus, the default is a small prior
precision \code{alpha = 1} and base measures that guarantee the right support.
For \code{y_max < Inf}, we simply use \code{Uniform{0,...,y_max}}. Otherwise,
we use \code{Geom(pi_geom)}, where \code{pi_geom} is elicited by fixing
the probability of exceeding the maximum observed value, \code{P(y > max(y))},
which we set at 0.10. Recall that the DP posterior for the marginal CDF
of \code{y} combines the base measure with the empirical measure: in expectation,
the base measure has weight \code{alpha/(n + alpha) = 1/(n+1)} while the empirical
part has weight \code{n/(n + alpha) = n/(n+1)}. Thus, the base measure's contribution is small,
and matters most for data values that may exceed \code{max(y)}.
}
\note{
The location (intercept) and scale (\code{sigma_epsilon}) are
not identified. These quantities are used for a location-scale
adjustment (or parameter expansion) that substantially improves the initial
approximation of the marginal CDF of \code{z}, but are otherwise
not interpretable.
}
\keyword{internal}
