% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{BrentMethod}
\alias{BrentMethod}
\title{Brent's method for optimization}
\usage{
BrentMethod(a = 0, b, fcn, tol = .Machine$double.eps^0.25, ...)
}
\arguments{
\item{a}{lower limit for search}

\item{b}{upper limit for search}

\item{fcn}{function to maximum}

\item{tol}{tolerance level}
}
\value{
a list of
}
\description{
Implementation for Brent's algorithm for maximizing a univariate function over an interval.
}
