% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{computeTimeRemaining}
\alias{computeTimeRemaining}
\title{Estimate the remaining time in the algorithm}
\usage{
computeTimeRemaining(nsi, timer0, nsims, nprints = 1)
}
\arguments{
\item{nsi}{current iteration}

\item{timer0}{initial timer value from \code{proc.time()[3]}}

\item{nsims}{total number of simulations}

\item{nprints}{total number of printed updates}
}
\value{
estimate of remaining time
}
\description{
Estimate the remaining time in the algorithm
}
\keyword{internal}
