% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR_frequentist.R
\name{confint.lmstar}
\alias{confint.lmstar}
\title{Compute asymptotic confidence intervals for STAR linear regression}
\usage{
\method{confint}{lmstar}(object, j, level = 0.95, include_plot = TRUE)
}
\arguments{
\item{object}{Object of class "lmstar" as output by \code{\link{lm_star}}}

\item{j}{the scalar column index for the desired confidence interval}

\item{level}{confidence level; default is 0.95}

\item{include_plot}{logical; if TRUE, include a plot of the profile likelihood}
}
\value{
the upper and lower endpoints of the confidence interval
}
\description{
For a linear regression model within the STAR framework,
compute (asymptotic) confidence intervals for a regression coefficient of interest.
Confidence intervals are computed by inverting the likelihood ratio test and
profiling the log-likelihood.
}
\examples{
# Simulate data with count-valued response y:
sim_dat = simulate_nb_lm(n = 100, p = 2)
y = sim_dat$y; X = sim_dat$X

# Select a transformation:
transformation = 'np'

#Estimate model
fit = lm_star(y, X, transformation)

# Confidence interval for the intercept:
ci_beta_0 = confint(fit, j = 1)
ci_beta_0

# Confidence interval for the slope:
ci_beta_1 = confint(fit, j = 2)
ci_beta_1

}
