% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expectation2_gRcpp}
\alias{expectation2_gRcpp}
\title{Compute E(Y^2) for a STAR process}
\usage{
expectation2_gRcpp(g_a_j, g_a_jp1, mu, sigma, Jmax)
}
\arguments{
\item{g_a_j}{\code{Jmax x 1} vector of g(a(j))}

\item{g_a_jp1}{\code{Jmax x 1} vector of g(a(j + 1))}

\item{mu}{\code{m x 1} vector of conditional expectations}

\item{sigma}{\code{m x 1} vector of conditional standard deviations}

\item{Jmax}{\code{m x 1} vector of maximum integer values to consider}
}
\value{
y2_hat \code{m x 1} vector of conditional expectations
}
\description{
Compute the conditional expectation of Y^2 for a STAR process Y
under a generic link function g.
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
