% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expectation_identity}
\alias{expectation_identity}
\title{Estimate the mean for a STAR process}
\usage{
expectation_identity(a, Jmax, Mu, sigma_t, Offset)
}
\arguments{
\item{a}{\code{Jmaxmax}-dimensonal vector of STAR integers a_j}

\item{Jmax}{\code{T x m} matrix of maximum integer values to consider}

\item{Mu}{\code{T x m} matrix of latent means}

\item{sigma_t}{\code{T}-dimensional vector of time-dependent latent error sd's}

\item{Offset}{\code{T x m} matrix of offsets}
}
\value{
Zhat \code{T x m} matrix of conditional expectations
}
\description{
Estimate the conditional expectation for a STAR process
under the identity link function.
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
\keyword{internal}
