% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{g_aj}
\alias{g_aj}
\title{Inverse rounding and transformation function}
\usage{
g_aj(y, transformation = "log", lambda = NULL, y_max = Inf)
}
\arguments{
\item{y}{argument(s) at which to evaluate \code{g(aj())}}

\item{transformation}{transformation to use for the latent process; must be one of
\itemize{
\item "identity" (identity transformation)
\item "log" (log transformation)
\item "sqrt" (square root transformation)
\item "box-cox" (box-cox transformation)
}}

\item{lambda}{the nonlinear parameter for the Box-Cox transformation}

\item{y_max}{a fixed and known upper bound for all observations; default is \code{Inf}}
}
\value{
\code{g(aj(y))}, where \code{g} and \code{aj} are determined by the
transformation
}
\description{
Compute the inverse of the rounding operator and transform, which
is used in the likelihood.
}
