% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{g_bnp_sparse_means}
\alias{g_bnp_sparse_means}
\title{Bayesian bootstrap-based transformation for sparse means}
\usage{
g_bnp_sparse_means(
  y,
  psi,
  pi_inc,
  zgrid = NULL,
  sigma_epsilon = 1,
  approx_Fz = FALSE
)
}
\arguments{
\item{y}{\code{n x 1} vector of observed counts}

\item{psi}{prior variance for the slab component}

\item{pi_inc}{prior inclusion  probability}

\item{zgrid}{optional vector of grid points for evaluating the CDF
of z (\code{Fz})}

\item{sigma_epsilon}{latent standard deviation; set to one for identifiability}

\item{approx_Fz}{logical; if TRUE, use a normal approximation for \code{Fz},
the marginal CDF of the latent z, which is faster and more stable}
}
\value{
A smooth monotone function which can be used for evaluations of the transformation
at each posterior draw.
}
\description{
Compute one posterior draw from the smoothed transformation
implied by (separate) Bayesian bootstrap models for the CDFs
of \code{y} and \code{X}.
This function is for the special case of the sparse means model.
}
