% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{g_deriv_approx}
\alias{g_deriv_approx}
\title{Approximate derivative of transformation}
\usage{
g_deriv_approx(g, t_grid)
}
\arguments{
\item{g}{the transformation function}

\item{t_grid}{grid of arguments at which to evaluate the transformation function}
}
\value{
A function which can be used for evaluations of the
(approximate) derivative of the transformation function.
}
\description{
Compute the (approximate) derivative of a transformation \code{g} based
on a smoothed grid evaluation.
}
\examples{
# Sample some data:
y = rpois(n = 500, lambda = 5)

# Empirical CDF transformation:
g_np = g_cdf(y, distribution = 'np')

# Grid for approximation:
t_grid = seq(1, max(y), length.out = 100)

# Compute the approximate derivative:
g_deriv = g_deriv_approx(g = g_np, t_grid = t_grid)
# Plot the transformation and then the derivative:
plot(t_grid, g_np(t_grid))
plot(t_grid, g_deriv(t_grid))

}
