% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{g_inv}
\alias{g_inv}
\title{Inverse transformation}
\usage{
g_inv(Fz_eval, Fy_eval, y_grid)
}
\arguments{
\item{Fz_eval}{the marginal CDF of z evaluated on some inputs}

\item{Fy_eval}{the marginal CDF of y evaluated on \code{y_grid}}

\item{y_grid}{a grid of non-negative integers}
}
\value{
The inverse transformation function evaluated on the same inputs
as \code{Fz_eval}.
}
\description{
Compute the inverse transformation on a vector of real-valued inputs
based on the marginal CDF of z and the marginal CDF of y.
}
\note{
The inputs for \code{Fz_eval} do not need to be known to compute
the inverse transformation, so they are not required for this function.

The function will return \code{NA} if \code{Fz_eval} is greater than all
\code{Fy_eval} values. When \code{y_max < Inf}, this implies that the inverse
is \code{y_max}; otherwise, it means that \code{y_grid} needs to use
larger values. Both are handled externally.
}
