% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{g_wcdf}
\alias{g_wcdf}
\title{Weighted cumulative distribution function (CDF)-based transformation}
\usage{
g_wcdf(y, distribution = "np", weights = NULL)
}
\arguments{
\item{y}{\code{n x 1} vector of observed counts}

\item{distribution}{the distribution used for the CDF; must be one of
\itemize{
\item "np" (empirical CDF)
\item "pois" (moment-matched marginal Poisson CDF)
\item "neg-bin" (moment-matched marginal Negative Binomial CDF)
}}

\item{weights}{an optional vector of weights}
}
\value{
A smooth monotone function which can be used for evaluations of the transformation.
}
\description{
Compute a CDF-based transformation using the observed count data.
The CDF can be estimated nonparametrically or parametrically based on the
Poisson or Negative-Binimial distributions. In the parametric case,
the parameters are determined based on the moments of \code{y}.
This function incorporates positive weights to determine the CDFs.
}
\examples{
# Sample some data:
y = rpois(n = 500, lambda = 5)
# And some weights:
w = runif(n = 500, min = 0, max = 10)

# Empirical CDF version:
g_np = g_wcdf(y, distribution = 'np', weights = w)

# Poisson version:
g_pois = g_wcdf(y, distribution = 'pois', weights = w)

# Negative binomial version:
g_negbin = g_wcdf(y, distribution = 'neg-bin', weights = w)

# Plot together:
t = 1:max(y) # grid
plot(t, g_np(t), type='l')
lines(t, g_pois(t), lty = 2)
lines(t, g_negbin(t), lty = 3)

}
