% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_MCMC.R
\name{Gauss_MCMC}
\alias{Gauss_MCMC}
\title{MCMC Algorithm for conditional Gaussian likelihood}
\usage{
Gauss_MCMC(
  y,
  sample_params,
  init_params,
  nsave = 5000,
  nburn = 5000,
  nskip = 2,
  verbose = TRUE
)
}
\arguments{
\item{y}{\code{n x 1} vector of observations}

\item{sample_params}{a function that inputs data \code{y} and a named list \code{params} containing
\enumerate{
\item \code{mu} \code{n x 1} vector of conditional means (fitted values)
\item \code{sigma} the conditional standard deviation
\item \code{coefficients} a named list of parameters that determine \code{mu}
}
and outputs an updated list \code{params} of samples from the full conditional posterior
distribution of \code{coefficients} and \code{sigma} (and updates \code{mu})}

\item{init_params}{an initializing function that inputs data \code{y}
and initializes the named list \code{params} of \code{mu}, \code{sigma}, and \code{coefficients}}

\item{nsave}{number of MCMC iterations to save}

\item{nburn}{number of MCMC iterations to discard}

\item{nskip}{number of MCMC iterations to skip between saving iterations,
i.e., save every (nskip + 1)th draw}

\item{verbose}{logical; if TRUE, print time remaining}
}
\value{
a list with the following elements:
\itemize{
\item \code{coefficients} the posterior mean of the coefficients
\item \code{fitted.values} the posterior mean of the conditional expectation of the data \code{y}
\item \code{post.coefficients} \code{nsave} posterior draws of the coefficients
\item \code{post.fitted.values} \code{nsave} posterior draws of the conditional mean of \code{y}
\item \code{post.pred} \code{nsave} draws from the posterior predictive distribution of \code{y}
\item \code{post.sigma} \code{nsave} draws from the posterior distribution of \code{sigma}
\item \code{post.log.like.point} \code{nsave} draws of the log-likelihood for each of the \code{n} observations
\item \code{logLik} the log-likelihood evaluated at the posterior means
\item \code{WAIC} Widely-Applicable/Watanabe-Akaike Information Criterion
\item \code{p_waic} Effective number of parameters based on WAIC
}
}
\description{
Run the MCMC algorithm for a conditional Gaussian likelihood given
(i) a function to initialize model parameters and
(ii) a function to sample (i.e., update) model parameters.
This is similar to the STAR framework, but without the transformation and rounding.
}
\examples{
# Fixme

}
