% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{h_ginv}
\alias{h_ginv}
\title{Rounding and inverse transformation function}
\usage{
h_ginv(z, transformation = "log", lambda = NULL, y_max = Inf)
}
\arguments{
\item{z}{argument(s) at which to evaluate \code{h(ginv())}}

\item{transformation}{transformation to use for the latent process; must be one of
\itemize{
\item "identity" (identity transformation)
\item "log" (log transformation)
\item "sqrt" (square root transformation)
\item "box-cox" (box-cox transformation)
}}

\item{lambda}{the nonlinear parameter for the Box-Cox transformation}

\item{y_max}{a fixed and known upper bound for all observations; default is \code{Inf}}
}
\value{
\code{h(ginv(z))}, where \code{h} and \code{ginv} are determined by the
transformation
}
\description{
Compute the inverse transformation and round, as defined in the data-generating
process. This is useful for computing posterior/prior predictive distributions.
}
