% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{HPDregion}
\alias{HPDregion}
\title{Compute highest posterior density (HPD) regions}
\usage{
HPDregion(post_pred, prob = 0.95, merge_gaps = FALSE)
}
\arguments{
\item{post_pred}{vector of draws from the posterior (predictive) distribution}

\item{prob}{numeric scalar in (0,1) giving the target probability content of the region}

\item{merge_gaps}{logical; if TRUE, add singleton points that are skipped over}
}
\value{
An ordered vector of values comprising the HPD region
}
\description{
Given a vector of draws from the posterior (predictive) distribution,
compute a \code{prob}% HPD region; i.e., the smallest set of values for which
their cumulative probability exceeds \code{prob}.
}
\note{
The HPD region is not necessarily contiguous. This function is primarily
designed for discrete distributions, so that the unique values in \code{post_pred}
accumulate multiple realizations. \code{merge_gaps} allows some smoothing over
"skipped" points, e.g., \code{{0,1,3}} becomes \code{{0,1,2,3}}.
}
