% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_MCMC.R
\name{init_params_lm_hs}
\alias{init_params_lm_hs}
\title{Initialize the parameters for a linear regression}
\usage{
init_params_lm_hs(y, X)
}
\arguments{
\item{y}{\code{n x 1} vector of data}

\item{X}{\code{n x p} matrix of predictors}
}
\value{
a named list \code{params} containing
\enumerate{
\item \code{mu} \code{n x 1} vector of conditional means (fitted values)
\item \code{sigma} the conditional standard deviation
\item \code{coefficients} a named list of parameters that determine \code{mu}
}
}
\description{
Initialize the parameters for a linear regression model assuming a
horseshoe prior for the (non-intercept) coefficients. The number of predictors
\code{p} may exceed the number of observations \code{n}.
}
\note{
The parameters in \code{coefficients} are:
\itemize{
\item \code{beta}: the \code{p x 1} vector of regression coefficients
\item \code{sigma_beta}: the \code{p x 1} vector of regression coefficient standard deviations
(local scale parameters)
\item \code{xi_sigma_beta}: the \code{p x 1} vector of parameter-expansion variables for \code{sigma_beta}
\item \code{lambda_beta}: the global scale parameter
\item \code{xi_lambda_beta}: the parameter-expansion variable for \code{lambda_beta}
components of \code{beta}
}
}
\examples{
# Simulate data for illustration:
sim_dat = simulate_nb_lm(n = 100, p = 5)
y = sim_dat$y; X = sim_dat$X

# Initialize:
params = init_params_lm_hs(y = y, X = X)
names(params)
names(params$coefficients)

}
