% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_MCMC.R
\name{init_params_mean}
\alias{init_params_mean}
\title{Initialize the parameters for a simple mean-only model}
\usage{
init_params_mean(y)
}
\arguments{
\item{y}{\code{n x 1} vector of data}
}
\value{
a named list \code{params} containing
\enumerate{
\item \code{mu}: vector of conditional means (fitted values)
\item \code{sigma}: the conditional standard deviation
\item \code{coefficients}: a named list of parameters that determine \code{mu}
}
}
\description{
Initialize the parameters for the model y ~ N(mu0, sigma^2)
with a flat prior on mu0.
}
\note{
The only parameter in \code{coefficients} is \code{mu0}.
Although redundant here, this parametrization is useful in other functions.
}
\examples{
# Example:
params = init_params_mean(y = 1:10)

}
