% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{invlogit}
\alias{invlogit}
\title{Compute the inverse log-odds}
\usage{
invlogit(x)
}
\arguments{
\item{x}{scalar or vector for which to compute the (componentwise) inverse log-odds}
}
\value{
A scalar or vector of values in (0,1)
}
\description{
Compute the inverse log-odds
}
\keyword{internal}
