% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikeRcpp}
\alias{logLikeRcpp}
\title{Compute the log-likelihood for STAR}
\usage{
logLikeRcpp(g_a_j, g_a_jp1, mu, sigma)
}
\arguments{
\item{g_a_j}{\code{m x 1} vector of g(a(j))}

\item{g_a_jp1}{\code{m x 1} vector of g(a(j + 1))}

\item{mu}{\code{m x 1} vector of conditional expectations}

\item{sigma}{\code{m x 1} vector of conditional standard deviations}
}
\value{
loglike scalar log-likelihood value
}
\description{
Compute the log-likelihood for a STAR model. The code here assumes
that the transformed real-valued process (z_star) has conditionally independent
components with means \code{mu} and standard deviations \code{sigma}.
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
