% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pmaxRcpp}
\alias{pmaxRcpp}
\title{pmax() in Rcpp}
\usage{
pmaxRcpp(v1, v2)
}
\arguments{
\item{v1}{\code{m x 1} vector}

\item{v2}{\code{m x 1} vector}
}
\value{
vm \code{m x 1} vector of pointwise maxima
}
\description{
Compute the pointwise max for two vectors of equal length
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
\keyword{internal}
