% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rdir}
\alias{rdir}
\title{Dirichlet sampler}
\usage{
rdir(params)
}
\arguments{
\item{params}{the vector of (nonnegative) Dirichlet parameters}
}
\value{
the vector of the Dirichlet draw on the simplex
}
\description{
Compute one Monte Carlo draw from a Dirichlet distribution
}
\details{
This function computes one draw from a Dirichlet distribution.
The output is on a simplex (nonnegative, sum to one) and has the same
length as the input \code{params}.
}
\examples{
# Example draw:
rdir(params = c(1,2,3))

}
