% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roaches}
\alias{roaches}
\title{Data on the efficacy of a pest management system at reducing the number of
roaches in urban apartments.}
\format{
## `roaches`
A data frame with 262 obs. of 6 variables:
\describe{
\item{y}{Number of roaches caught}
\item{roach1}{Pretreatment number of roaches}
\item{treatment}{Treatment indicator}
\item{senior}{ Indicator for only elderly residents in building}
\item{exposure2}{Number of days for which the roach traps were used}
}
}
\source{
Gelman and Hill (2007); package `rstanarm`
}
\usage{
roaches
}
\description{
Data on the efficacy of a pest management system at reducing the number of
roaches in urban apartments.
}
\keyword{datasets}
