% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtruncnormRcpp}
\alias{rtruncnormRcpp}
\title{Sample from a truncated normal distribution}
\usage{
rtruncnormRcpp(y_lower, y_upper, mu, sigma, u_rand)
}
\arguments{
\item{y_lower}{\code{m x 1} vector of lower endpoints}

\item{y_upper}{\code{m x 1} vector of upper endpoints}

\item{mu}{\code{m x 1} vector of conditional expectations}

\item{sigma}{\code{m x 1} vector of conditional standard deviations}

\item{u_rand}{\code{m x 1} vector of uniform random variables}
}
\value{
z_star \code{m x 1} draw from the truncated normal distribution
}
\description{
Sample from a truncated normal distribution. Samples are drawn
componentwise, so each component of the vector is allowed its own
mean, standard deviation, and upper and lower limits. The components
are assumed to be independent.
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
\keyword{internal}
