% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{sample_bam_orthog}
\alias{sample_bam_orthog}
\title{Sample the parameters for an additive model}
\usage{
sample_bam_orthog(
  y,
  X_lin,
  X_nonlin,
  params,
  A = 10^4,
  B_all = NULL,
  diagBtB_all = NULL,
  XtX = NULL
)
}
\arguments{
\item{y}{\code{n x 1} vector of data}

\item{X_lin}{\code{n x pL} matrix of predictors to be modelled as linear}

\item{X_nonlin}{\code{n x pNL} matrix of predictors to be modelled as nonlinear}

\item{params}{the named list of parameters containing
\enumerate{
\item \code{mu}: vector of conditional means (fitted values)
\item \code{sigma}: the conditional standard deviation
\item \code{coefficients}: a named list of parameters that determine \code{mu}
}}

\item{A}{the prior scale for \code{sigma_beta}, which we assume follows a Uniform(0, A) prior.}

\item{B_all}{optional \code{pNL}-dimensional list of \code{n x L[j]} dimensional
basis matrices for each nonlinear term j=1,...,pNL; if NULL, compute internally}

\item{diagBtB_all}{optional \code{pNL}-dimensional list of \code{diag(crossprod(B_all[[j]]))};
if NULL, compute internally}

\item{XtX}{optional \code{p x p} matrix of \code{crossprod(X)} (one-time cost);
if NULL, compute internally}
}
\value{
The updated named list \code{params} with draws from the full conditional distributions
of \code{sigma} and \code{coefficients} (and updated \code{mu}).
}
\description{
Sample the parameters for an additive model, which may contain
both linear and nonlinear predictors. The nonlinear terms are modeled
using orthogonalized splines. The sampler draws the linear terms
jointly and then samples each vector of nonlinear coefficients using
Bayesian backfitting (i.e., conditional on all other nonlinear and linear terms).
}
\note{
The parameters in \code{coefficients} are:
\itemize{
\item \code{beta}: the \code{p x 1} linear coefficients, including the linear terms from \code{X_nonlin}
\item \code{f_j}: the \code{n x pNL} matrix of fitted values for each nonlinear function
\item \code{theta_j}: the \code{pNL}-dimensional of nonlinear basis coefficients
\item \code{sigma_beta}: \code{p x 1} vector of linear regression coefficient standard deviations
\item \code{sigma_theta_j}: \code{pNL x 1} vector of nonlinear coefficient standard deviations
}
}
\keyword{internal}
