% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{sample_lm_ridge}
\alias{sample_lm_ridge}
\title{Sample linear regressionparameters assuming a ridge prior}
\usage{
sample_lm_ridge(y, X, params, A = 10^4, XtX = NULL)
}
\arguments{
\item{y}{\code{n x 1} vector of data}

\item{X}{\code{n x p} matrix of predictors}

\item{params}{the named list of parameters containing
\enumerate{
\item \code{mu}: vector of conditional means (fitted values)
\item \code{sigma}: the conditional standard deviation
\item \code{coefficients}: a named list of parameters that determine \code{mu}
}}

\item{A}{the prior scale for \code{sigma_beta}, which we assume follows a Uniform(0, A) prior.}

\item{XtX}{the \code{p x p} matrix of \code{crossprod(X)} (one-time cost);
if NULL, compute within the function}
}
\value{
The updated named list \code{params} with draws from the full conditional distributions
of \code{sigma} and \code{coefficients} (and updated \code{mu}).
}
\description{
Sample the parameters for a linear regression model assuming a
ridge prior for the (non-intercept) coefficients. The number of predictors
\code{p} may exceed the number of observations \code{n}.
}
\note{
The parameters in \code{coefficients} are:
\itemize{
\item \code{beta}: the \code{p x 1} vector of regression coefficients
\item \code{sigma_beta}: the prior standard deviation for the (non-intercept)
components of \code{beta}
}
}
\examples{
# Simulate data for illustration:
sim_dat = simulate_nb_lm(n = 100, p = 5)
y = sim_dat$y; X = sim_dat$X

# Initialize:
params = init_params_lm(y = y, X = X)

# Sample:
params = sample_params_lm(y = y, X = X, params = params)
names(params)
names(params$coefficients)

}
\keyword{internal}
