% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{simulate_nb_lm}
\alias{simulate_nb_lm}
\title{Simulate count data from a linear regression}
\usage{
simulate_nb_lm(
  n = 100,
  p = 10,
  r_nb = 1,
  b_int = log(1.5),
  b_sig = log(2),
  sigma_true = sqrt(2 * log(1)),
  ar1 = 0,
  intercept = FALSE
)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of predictors (including the intercept)}

\item{r_nb}{the dispersion parameter of the Negative Binomial dispersion;
smaller values imply greater overdispersion, while larger values approximate the Poisson distribution.}

\item{b_int}{intercept; default is log(1.5), which implies the expected count is 1.5
when all predictors are zero}

\item{b_sig}{regression coefficients for true signals; default is log(2.0), which implies a
twofold increase in the expected counts for a one unit increase in x}

\item{sigma_true}{standard deviation of the Gaussian innovation; default is zero.}

\item{ar1}{the autoregressive coefficient among the columns of the X matrix; default is zero.}

\item{intercept}{a Boolean indicating whether an intercept column should be included
in the returned design matrix; default is FALSE}
}
\value{
A named list with the simulated count response \code{y}, the simulated design matrix \code{X}
(possibly including the intercept), the true expected counts \code{Ey},
and the true regression coefficients \code{beta_true}.
}
\description{
Simulate data from a negative-binomial distribution with linear mean function.
}
\details{
The log-expected counts are modeled as a linear function of covariates, possibly
with additional Gaussian noise (on the log-scale). We assume that half of the predictors
are associated with the response, i.e., true signals. For sufficiently large dispersion
parameter \code{r_nb}, the distribution will approximate a Poisson distribution.
Here, the predictor variables are simulated from independent standard normal distributions.
}
\note{
Specifying \code{sigma_true = sqrt(2*log(1 + a))} implies that the expected counts are
inflated by \code{100*a}\% (relative to \code{exp(X*beta)}), in addition to providing additional
overdispersion.
}
\examples{
# Simulate and plot the count data:
sim_dat = simulate_nb_lm(n = 100, p = 10);
plot(sim_dat$y)

}
