% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{spline_star_gibbs_bnp}
\alias{spline_star_gibbs_bnp}
\title{Gibbs sampler for STAR spline regression with BNP transformation}
\usage{
spline_star_gibbs_bnp(
  y,
  x = NULL,
  x_test = NULL,
  y_max = Inf,
  psi = NULL,
  nbasis = NULL,
  alpha = 1,
  P0 = NULL,
  nsave = 1000,
  nburn = 1000,
  nskip = 0,
  verbose = TRUE
)
}
\arguments{
\item{y}{\code{n x 1} vector of observed counts}

\item{x}{\code{n x 1} vector of observation points; if NULL, assume equally-spaced on [0,1]}

\item{x_test}{\code{n_test x 1} vector of testing points; default is \code{x}}

\item{y_max}{a fixed and known upper bound for all observations; default is \code{Inf}}

\item{psi}{prior variance (inverse smoothing parameter); if NULL,
sample this parameter}

\item{nbasis}{number of spline basis functions; if NULL, use the default from \code{spikeSlabGAM::sm}}

\item{alpha}{prior precision for the Dirichlet Process prior; default is one}

\item{P0}{function to evaluate the base measure PMF supported on \code{{0,...,y_max}};
see below for default values when unspecified (\code{NULL})}

\item{nsave}{number of MCMC iterations to save}

\item{nburn}{number of MCMC iterations to discard}

\item{nskip}{number of MCMC iterations to skip between saving iterations,
i.e., save every (nskip + 1)th draw}
}
\value{
a list with the following elements:
\itemize{
\item \code{coefficients}: the posterior mean of the spline coefficients
\item \code{fitted.values} the posterior predictive mean at the test points \code{x_test}
\item \code{post.beta}: \code{nsave x p} samples from the posterior distribution
of the regression coefficients
\item \code{post.pred}: \code{nsave x n_test} samples
from the posterior predictive distribution at test points \code{x_test}
\item \code{post.g}: \code{nsave} posterior samples of the transformation
evaluated at \code{1:max(y)}
\item \code{post.psi}: \code{nsave} draws from the prior variance (inverse smoothing) \code{psi}
}
}
\description{
Compute MCMC samples from the posterior and predictive
distributions of a STAR spline regression model with a
Bayesian nonparametric (BNP) transformation. Cubic B-splines are
used with a prior that penalizes roughness.
}
\details{
This function provides fully Bayesian inference for a
transformed spline model with discrete data. The sampling algorithm draws the
transformation \code{g} directly from its *marginal* posterior distribution
and then uses a Gibbs sampler for the latent data \code{z} and the
spline coefficients \code{beta}.

The BNP model for the transformation \code{g} is derived from a Dirichlet Process (DP)
prior on the marginal CDF of \code{y}. The user can specify the prior precision
\code{alpha} and the base measure, e.g., \code{P0 = function(t) dpois(t, lambda = 10)}.
Otherwise, the default approach views the base measure as a means to ensure
correct support and avoid boundary issues, both of which are concerns with the
Bayesian bootstrap (\code{alpha = 0}). Thus, the default is a small prior
precision \code{alpha = 1} and base measures that guarantee the right support.
For \code{y_max < Inf}, we simply use \code{Uniform{0,...,y_max}}. Otherwise,
we use \code{Geom(pi_geom)}, where \code{pi_geom} is elicited by fixing
the probability of exceeding the maximum observed value, \code{P(y > max(y))},
which we set at 0.10. Recall that the DP posterior for the marginal CDF
of \code{y} combines the base measure with the empirical measure: in expectation,
the base measure has weight \code{alpha/(n + alpha) = 1/(n+1)} while the empirical
part has weight \code{n/(n + alpha) = n/(n+1)}. Thus, the base measure's contribution is small,
and matters most for data values that may exceed \code{max(y)}.
}
\note{
The location (intercept) and scale (\code{sigma_epsilon}) are
not identified. These quantities are used for a location-scale
adjustment (or parameter expansion) that substantially improves the initial
approximation of the marginal CDF of \code{z}, but are otherwise
not interpretable.
}
\keyword{internal}
