% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_EM.R
\name{star_EM_wls}
\alias{star_EM_wls}
\title{EM Algorithm for the STAR linear model with weighted least squares}
\usage{
star_EM_wls(y, X, transformation = "log", lambda = NULL, y_max = Inf,
  weights = NULL, tol = 10^-10, max_iters = 1000)
}
\arguments{
\item{y}{\code{n x 1} vector of observed counts}

\item{X}{\code{n x p} matrix of predictors}

\item{transformation}{transformation to use for the latent process; must be one of
\itemize{
\item "identity" (identity transformation)
\item "log" (log transformation)
\item "sqrt" (square root transformation)
\item "box-cox" (box-cox transformation)
}}

\item{lambda}{the nonlinear parameter for the Box-Cox transformation; otherwise ignored}

\item{y_max}{a fixed and known upper bound for all observations; default is \code{Inf}}

\item{weights}{an optional vector of weights to be used in the fitting process, which
produces weighted least squares estimators.}

\item{tol}{tolerance for stopping the EM algorithm; default is 10^-10;}

\item{max_iters}{maximum number of EM iterations before stopping; default is 1000}
}
\value{
a list with the following elements:
\itemize{
\item \code{coefficients} the MLEs of the coefficients
\item \code{fitted.values} the fitted values at the MLEs
\item \code{sigma.hat} the MLE of the standard deviation
\item \code{mu.hat} the MLE of the conditional mean (on the transformed scale)
\item \code{z.hat} the estimated latent variables (on the transformed scale) at the MLEs
\item \code{residuals} the Dunn-Smyth residuals
\item \code{logLik} the log-likelihood at the MLEs
\item and other parameters that
(i) track the parameters across EM iterations and
(ii) record the model specifications
}
}
\description{
Compute the MLEs and log-likelihood for the STAR linear model.
The regression coefficients are estimated using weighted least squares within
an EM algorithm. Standard function calls including \code{coefficients()},
\code{fitted()}, and \code{residuals()} apply.
}
\note{
For the Box-Cox transformation, a \code{NULL} value of
\code{lambda} requires estimation of \code{lambda}. The maximum likelihood
estimator is computed over a grid of values within the EM algorithm.
}
