% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_MCMC.R
\name{star_np_MCMC2}
\alias{star_np_MCMC2}
\title{MCMC sampler for STAR with nonparametric link function}
\usage{
star_np_MCMC2(y, sample_params, init_params, lambda_prior = 1/2,
  y_max = Inf, nsave = 5000, nburn = 5000, nskip = 2,
  save_y_hat = FALSE, verbose = TRUE)
}
\arguments{
\item{y}{\code{n x 1} vector of observed counts}

\item{sample_params}{a function that inputs data \code{y} and a named list \code{params} containing
\enumerate{
\item \code{mu}: vector of conditional means (fitted values)
\item \code{sigma}: the conditional standard deviation
\item \code{coefficients}: a named list of parameters that determine \code{mu}
}
and outputs an updated list \code{params} of samples from the full conditional posterior
distribution of \code{coefficients} and \code{sigma} (and updates \code{mu})}

\item{init_params}{an initializing function that inputs data \code{y}
and initializes the named list \code{params} of \code{mu}, \code{sigma}, and \code{coefficients}}

\item{lambda_prior}{the prior mean for the transformation g() is the Box-Cox function with
parameter \code{lambda_prior}}

\item{y_max}{a fixed and known upper bound for all observations; default is \code{Inf}}

\item{nsave}{number of MCMC iterations to save}

\item{nburn}{number of MCMC iterations to discard}

\item{nskip}{number of MCMC iterations to skip between saving iterations,
i.e., save every (nskip + 1)th draw}

\item{save_y_hat}{logical; if TRUE, compute and save the posterior draws of
the expected counts, E(y), which may be slow to compute}

\item{verbose}{logical; if TRUE, print time remaining}
}
\value{
a list with the following elements:
\itemize{
\item \code{coefficients}: the posterior mean of the coefficients
\item \code{fitted.values}: the posterior mean of the conditional expectation of the counts \code{y}
\item \code{post.coefficients}: posterior draws of the coefficients
\item \code{post.fitted.values}: posterior draws of the conditional mean of the counts \code{y}
\item \code{post.pred}: draws from the posterior predictive distribution of \code{y}
\item \code{post.g}: draws from the posterior distribution of the transformation \code{g}
\item \code{post.sigma}: draws from the posterior distribution of \code{sigma}
\item \code{post.alpha}: draws from the posterior distribution of \code{alpha},
the prior concentration of the transformation \code{g} coefficients
\item \code{post.log.like.point}: draws of the log-likelihood for each of the \code{n} observations
\item \code{WAIC}: Widely-Applicable/Watanabe-Akaike Information Criterion
\item \code{p_waic}: Effective number of parameters based on WAIC
}
}
\description{
Run the MCMC algorithm for STAR given
\enumerate{
\item a function to initialize model parameters; and
\item a function to sample (i.e., update) model parameters.
}
The transformation is modeled as an unknown, monotone function
using I-splines. A data augmentation step is used for posterior sampling.
}
