% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_MCMC.R
\name{STAR_sparse_means}
\alias{STAR_sparse_means}
\title{Stochastic search for the STAR sparse means model}
\usage{
STAR_sparse_means(
  y,
  transformation = "identity",
  y_min = -Inf,
  y_max = Inf,
  psi = NULL,
  a_pi = 1,
  b_pi = 1,
  approx_Fz = FALSE,
  approx_Fy = FALSE,
  nsave = 1000,
  nburn = 1000,
  nskip = 0,
  verbose = TRUE
)
}
\arguments{
\item{y}{\code{n x 1} vector of integers}

\item{transformation}{transformation to use for the latent data; must be one of
\itemize{
\item "identity" (signed identity transformation)
\item "sqrt" (signed square root transformation)
\item "bnp" (Bayesian nonparametric transformation using the Bayesian bootstrap)
\item "np" (nonparametric transformation estimated from empirical CDF)
\item "pois" (transformation for moment-matched marginal Poisson CDF)
\item "neg-bin" (transformation for moment-matched marginal Negative Binomial CDF)
}}

\item{y_min}{a fixed and known upper bound for all observations; default is \code{-Inf}}

\item{y_max}{a fixed and known upper bound for all observations; default is \code{Inf}}

\item{psi}{prior variance for the slab component;
if NULL, assume a Unif(0, n) prior}

\item{a_pi}{prior shape1 parameter for the inclusion probability;
default is 1 for uniform}

\item{b_pi}{prior shape2 parameter for the inclusion probability;
#' default is 1 for uniform}

\item{approx_Fz}{logical; in BNP transformation, apply a (fast and stable)
normal approximation for the marginal CDF of the latent data}

\item{approx_Fy}{logical; in BNP transformation, approximate
the marginal CDF of \code{y} using the empirical CDF}

\item{nsave}{number of MCMC iterations to save}

\item{nburn}{number of MCMC iterations to discard}

\item{nskip}{number of MCMC iterations to skip between saving iterations,
i.e., save every (nskip + 1)th draw}

\item{verbose}{logical; if TRUE, print time remaining}
}
\value{
a list with the following elements:
\itemize{
\item \code{post_gamma}: \code{nsave x n} samples from the posterior distribution
of the inclusion indicators
\item \code{post_pi}: \code{nsave} samples from the posterior distribution
of the inclusion probability
\item \code{post_psi}: \code{nsave} samples from the posterior distribution
of the prior variance
\item \code{post_theta}: \code{nsave} samples from the posterior distribution
of the regression coefficients
\item \code{post_g}: \code{nsave} posterior samples of the transformation
evaluated at the unique \code{y} values (only applies for 'bnp' transformations)
}
}
\description{
Compute Gibbs samples from the posterior distribution
of the inclusion indicators for the sparse means model.
The inclusion probability is assigned a Beta(a_pi, b_pi) prior
and is learned as well.
}
\details{
STAR defines an integer-valued probability model by
(1) specifying a Gaussian model for continuous *latent* data and
(2) connecting the latent data to the observed data via a
*transformation and rounding* operation. Here, the continuous
latent data model is a sparse normal means model of the form
z_i = theta_i + epsilon_i with a spike-and-slab prior on
theta_i.

There are several options for the transformation. First, the transformation
can belong to the signed *Box-Cox* family, which includes the known transformations
'identity' and 'sqrt'. Second, the transformation
can be estimated (before model fitting) using the empirical distribution of the
data \code{y}. Options in this case include the empirical cumulative
distribution function (CDF), which is fully nonparametric ('np'), or the parametric
alternatives based on Poisson ('pois') or Negative-Binomial ('neg-bin')
distributions. For the parametric distributions, the parameters of the distribution
are estimated using moments (means and variances) of \code{y}. The distribution-based
transformations approximately preserve the mean and variance of the count data \code{y}
on the latent data scale, which lends interpretability to the model parameters.
Lastly, the transformation can be modeled using the Bayesian bootstrap ('bnp'),
which is a Bayesian nonparametric model and incorporates the uncertainty
about the transformation into posterior and predictive inference.

There are several options for the prior variance \code{psi}.
First, it can be specified directly. Second, it can be assigned
a Uniform(0,n) prior and sampled within the MCMC.
}
\examples{
# Simulate some data:
y = round(c(rnorm(n = 100, mean = 0),
            rnorm(n = 100, mean = 2)))

# Fit the model:
fit = STAR_sparse_means(y, nsave = 100, nburn = 100) # for a quick example
names(fit)

# Posterior inclusion probabilities:
pip = colMeans(fit$post_gamma)
plot(pip, y)

# Check the MCMC efficiency:
getEffSize(fit$post_theta) # coefficients

}
