% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{truncnorm_mom}
\alias{truncnorm_mom}
\title{Compute the first and second moment of a truncated normal}
\usage{
truncnorm_mom(a, b, mu, sig)
}
\arguments{
\item{a}{lower endpoint}

\item{b}{upper endpoint}

\item{mu}{expected value of the non-truncated normal distribution}

\item{sig}{standard deviation of the non-truncated normal distribution}
}
\value{
a list containing the first moment \code{m1} and the second moment \code{m2}
}
\description{
Given lower and upper endpoints and the mean and standard deviation
of a (non-truncated) normal distribution, compute the first and second
moment of the truncated normal distribution. All inputs may be scalars
or vectors.
}
\keyword{internal}
