% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warpDLM.R
\name{update_struct}
\alias{update_struct}
\title{Update parameters for warpDLM model with trend DLM}
\usage{
update_struct(fit, z_star, theta)
}
\arguments{
\item{fit}{the KFAS model object describing the DLM}

\item{z_star}{the latest draw of z*}

\item{theta}{the latest draw of the latent state(s) theta}
}
\value{
A KFAS model object (of class SSModel) updated with the newly
sampled variance parameters
}
\description{
This function serves to update the warpDLM variance parameters
when the underlying DLM is a structural model (i.e. local level
or local linear trend). It assumes a Unif(0,A=10^4) prior on all
standard deviations.
}
\keyword{internal}
