% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR_Bayesian.R
\name{blm_star}
\alias{blm_star}
\title{STAR Bayesian Linear Regression}
\usage{
blm_star(
  y,
  X,
  X_test = X,
  transformation = "np",
  y_max = Inf,
  prior = "gprior",
  use_MCMC = TRUE,
  nsave = 5000,
  nburn = 5000,
  nskip = 0,
  method_sigma = "mle",
  approx_Fz = FALSE,
  approx_Fy = FALSE,
  psi = NULL,
  compute_marg = FALSE
)
}
\value{
a list with the following elements:
}
\description{
STAR Bayesian Linear Regression
}
