% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_EM.R
\name{randomForest_star}
\alias{randomForest_star}
\title{EM Algorithm for Random Forest STAR}
\usage{
randomForest_star(y, X, X.test = NULL, ntree = 500,
  mtry = max(floor(ncol(X)/3), 1), nodesize = 5,
  transformation = "log", lambda = NULL, y_max = Inf, tol = 10^-6,
  max_iters = 500)
}
\arguments{
\item{y}{\code{n x 1} vector of observed counts}

\item{X}{\code{n x p} matrix of predictors}

\item{X.test}{\code{m x p} matrix of out-of-sample predictors}

\item{ntree}{Number of trees to grow.
This should not be set to too small a number, to ensure that every input row gets predicted at least a few times.
Default is 200.}

\item{mtry}{Number of variables randomly sampled as candidates at each split.
Default is p/3.}

\item{nodesize}{Minimum size of terminal nodes. Setting this number larger causes smaller trees to be grown (and thus take less time).
Default is 5.}

\item{transformation}{transformation to use for the latent process; must be one of
\itemize{
\item "identity" (identity transformation)
\item "log" (log transformation)
\item "sqrt" (square root transformation)
\item "box-cox" (box-cox transformation)
}}

\item{lambda}{the nonlinear parameter for the Box-Cox transformation; otherwise ignored}

\item{y_max}{a fixed and known upper bound for all observations; default is \code{Inf}}

\item{tol}{tolerance for stopping the EM algorithm; default is 10^-10;}

\item{max_iters}{maximum number of EM iterations before stopping; default is 1000}
}
\value{
a list with the following elements:
\itemize{
\item \code{fitted.values}: the fitted values at the MLEs based on out-of-bag samples (training)
\item \code{fitted.values.test}: the fitted values at the MLEs (testing)
\item \code{sigma.hat}: the MLE of the standard deviation
\item \code{mu.hat}: the MLE of the conditional mean (on the transformed scale)
\item \code{z.hat}: the estimated latent variables (on the transformed scale) at the MLEs
\item \code{residuals}: the Dunn-Smyth residuals
\item \code{logLik}: the log-likelihood at the MLEs based on out-of-bag samples
\item \code{rfObj}: the object returned by randomForest() at the MLEs
\item and other parameters that
(i) track the parameters across EM iterations and
(ii) record the model specifications
}
}
\description{
Compute the MLEs and log-likelihood for the Random Forest STAR model. The STAR model requires
a transformation (such as log, sqrt, or Box-Cox) and will estimate the conditional mean
on the transformed scale using random forests. Standard function calls
including \code{fitted()} and \code{residuals()} apply.
}
\details{
For the Box-Cox transformation, a \code{NULL} value of
\code{lambda} requires estimation of \code{lambda}. The maximum likelihood
estimator is computed over a grid of values within the EM algorithm.

The fitted values are computed using out-of-bag samples. As a result,
the log-likelihood is based on out-of-bag prediction, and it is similarly
straightforward to compute out-of-bag squared and absolute errors.
}
\note{
Since the random foreset produces random predictions, the EM algorithm
will never converge exactly.
}
\examples{
\dontrun{
# Simulate data with count-valued response y:
sim_dat = simulate_nb_friedman(n = 100, p = 10)
y = sim_dat$y; X = sim_dat$X

# EM algorithm for STAR (using the log-link)
fit_em = randomForest_star(y = y, X = X,
                 transformation = 'log',
                 max_iters = 100)

# Fitted values (out-of-bag)
y_hat = fitted(fit_em)
plot(y_hat, y);

# Residuals:
plot(residuals(fit_em))
qqnorm(residuals(fit_em)); qqline(residuals(fit_em))

# Log-likelihood at MLEs (out-of-bag):
fit_em$logLik
}

}
